+++ matgl/config.py
@@ -2,11 +2,12 @@
 from __future__ import annotations
 
 import os
+import shutil
 from pathlib import Path
 
-import numpy as np
-import torch
-
+"""
+Default set of elements supported by universal matgl models.
+"""
 DEFAULT_ELEMENT_TYPES = (
     "H",
     "He",
@@ -99,56 +100,28 @@ DEFAULT_ELEMENT_TYPES = (
     "Pu",
 )
 
-DTYPES = {
-    "float32": {"numpy": np.float32, "torch": torch.float32},
-    "float16": {"numpy": np.float16, "torch": torch.float16},
-    "int32": {"numpy": np.int32, "torch": torch.int32},
-    "int16": {"numpy": np.int16, "torch": torch.int16},
-}
-
+"""
+Default location of the cache for matgl, e.g., for storing downloaded models.
+"""
 MATGL_CACHE = Path(os.path.expanduser("~")) / ".matgl"
+
+"""
+Download url for pre-trained models.
+"""
 PRETRAINED_MODELS_BASE_URL = "https://github.com/materialsvirtuallab/matgl/raw/main/pretrained_models/"
 
+"""
+This is an int representing a model version. It is mainly for detecting and warning about the use of old pre-trained
+models. This version number is different from the code version number because it depends on whether
+backward-incompatible architectural changes are made (which hopefully should be less often than regular code changes).
+"""
+MODEL_VERSION = 1
 
-class DataType:
+
+def clear_cache():
     """
-    Tensorflow and numpy data types. Used to choose between float16 and float32
+    Deletes all files in the matgl.cache. This is used to clean out downloaded models.
     """
-
-    # np_float = tf.keras.mixed_precision.global_policy().compute_dtype
-    np_float = "float32"
-    np_int = "int32"
-    # torch_float = tf.keras.mixed_precision.global_policy().compute_dtype
-    torch_float = torch.float32
-    torch_int = torch.int32
-
-    @classmethod
-    def set_dtype(cls, data_type: str) -> None:
-        """
-        Class method to set the data types
-        Args:
-            data_type (str): '16' or '32'
-        """
-        if data_type.endswith("32"):
-            float_key = "float32"
-            int_key = "int32"
-        elif data_type.endswith("16"):
-            float_key = "float16"
-            int_key = "int16"
-        else:
-            raise ValueError("Data type not known, choose '16' or '32'")
-
-        cls.np_float = DTYPES[float_key]["numpy"]  # type: ignore
-        cls.torch_float = DTYPES[float_key]["torch"]  # type: ignore
-        cls.np_int = DTYPES[int_key]["numpy"]  # type: ignore
-        cls.torch_int = DTYPES[int_key]["torch"]  # type: ignore
-
-
-def set_global_dtypes(data_type) -> None:
-    """
-    Function to set the data types
-    Args:
-        data_type (str): '16' or '32'
-    Returns:
-    """
-    DataType.set_dtype(data_type)
+    r = input(f"Do you really want to delete everything in {MATGL_CACHE} (y|n)? ")
+    if r.lower() == "y":
+        shutil.rmtree(MATGL_CACHE)