+++ matgl/models/__init__.py
@@ -4,5 +4,6 @@ Package containing model implementations.
 
 from __future__ import annotations
 
-from matgl.models._m3gnet import M3GNet
-from matgl.models._megnet import MEGNet
+from ._m3gnet import M3GNet
+from ._megnet import MEGNet
+from ._wrappers import TransformedTargetModel