+++ matgl/utils/io.py
@@ -7,12 +7,13 @@ import inspect
 import json
 import logging
 import os
+import warnings
 from pathlib import Path
 
 import requests
 import torch
 
-from matgl.config import MATGL_CACHE, PRETRAINED_MODELS_BASE_URL
+from matgl.config import MATGL_CACHE, MODEL_VERSION, PRETRAINED_MODELS_BASE_URL
 
 logger = logging.getLogger(__file__)
 
@@ -22,7 +23,7 @@ class IOMixIn:
     Mixin class for model saving and loading.
     """
 
-    def save_args(self, locals: dict, kwargs: dict) -> None:
+    def save_args(self, locals: dict, kwargs: dict | None = None) -> None:
         r"""
         Method to save args into a private _init_args variable. This should be called after super in the __init__
         method, e.g., `self.save_args(locals(), kwargs)`.
@@ -32,8 +33,20 @@ class IOMixIn:
             kwargs: kwargs passed to the class.
         """
         args = inspect.getfullargspec(self.__class__.__init__).args
-        self._init_args = {k: v for k, v in locals.items() if k in args and k not in ("self", "__class__")}
-        self._init_args.update(kwargs)
+        d = {k: v for k, v in locals.items() if k in args and k not in ("self", "__class__")}
+        if kwargs is not None:
+            d.update(kwargs)
+
+        # If one of the args is a subclass of IOMixIn, we will serialize that class.
+        for k, v in d.items():
+            if issubclass(v.__class__, IOMixIn):
+                d[k] = {
+                    "@class": v.__class__.__name__,
+                    "@module": v.__class__.__module__,
+                    "@model_version": MODEL_VERSION,
+                    "init_args": v._init_args,
+                }
+        self._init_args = d
 
     def save(self, path: str | Path = ".", metadata: dict | None = None, makedirs: bool = True):
         """
@@ -52,11 +65,16 @@ class IOMixIn:
         path = Path(path)
         if makedirs:
             os.makedirs(path, exist_ok=True)
+
         torch.save(self._init_args, path / "model.pt")  # type: ignore
         torch.save(self.state_dict(), path / "state.pt")  # type: ignore
-
-        # This json dump of model args is purely for ease of reference. It is not used to deserialize the model.
-        d = {"name": self.__class__.__name__, "metadata": metadata, "kwargs": self._init_args}  # type: ignore
+        d = {
+            "@class": self.__class__.__name__,
+            "@module": self.__class__.__module__,
+            "@model_version": MODEL_VERSION,
+            "metadata": metadata,
+            "kwargs": self._init_args,
+        }  # type: ignore
         with open(path / "model.json", "w") as f:
             json.dump(d, f, default=lambda o: str(o), indent=4)
 
@@ -66,9 +84,8 @@ class IOMixIn:
         Load the model weights from a directory.
 
         Args:
-            path (str|path): Path to saved model or name of pre-trained model. The search order is
-                path, followed by model name in PRETRAINED_MODELS_PATH, followed by download from
-                PRETRAINED_MODELS_BASE_URL.
+            path (str|path): Path to saved model or name of pre-trained model. The search order is path, followed by
+                download from PRETRAINED_MODELS_BASE_URL (with caching).
             include_json (bool): If True, the "model.json" file is also loaded. This file can contain metadata about
                 the model, e.g., if scaling was performed in training the model, this file may contain the mean and
                 standard deviation of the models, which needs to be applied to the final predictions.
@@ -85,8 +102,6 @@ class IOMixIn:
 
         if all((path / fn).exists() for fn in fnames):
             fpaths = {fn: path / fn for fn in fnames}
-        elif all((MATGL_CACHE / path / fn).exists() for fn in fnames):
-            fpaths = {fn: MATGL_CACHE / path / fn for fn in fnames}
         else:
             try:
                 fpaths = {
@@ -102,7 +117,18 @@ class IOMixIn:
             state = torch.load(fpaths["state.pt"], map_location=torch.device("cpu"))
         else:
             state = torch.load(fpaths["state.pt"])
-        model = cls(**torch.load(fpaths["model.pt"]))
+        d = torch.load(fpaths["model.pt"])
+
+        # Deserialize any args that are IOMixIn subclasses.
+        for k, v in d.items():
+            if isinstance(v, dict) and "@class" in v and "@module" in v:
+                modname = v["@module"]
+                classname = v["@class"]
+                mod = __import__(modname, globals(), locals(), [classname], 0)
+                cls_ = getattr(mod, classname)
+                d[k] = cls_(**v["init_args"])
+        d = {k: v for k, v in d.items() if not k.startswith("@")}
+        model = cls(**d)
         model.load_state_dict(state)  # type: ignore
 
         if include_json:
@@ -167,3 +193,50 @@ class RemoteFile:
             exc_tb: Usual meaning in __exit__.
         """
         self.stream.close()
+
+
+def load_model(path: Path, **kwargs):
+    r"""
+    Convenience method to load a model from a directory or name.
+
+    Args:
+        path (str|path): Path to saved model or name of pre-trained model. The search order is path, followed by
+            download from PRETRAINED_MODELS_BASE_URL (with caching).
+        **kwargs: Additional kwargs passed to RemoteFile class. E.g., a useful one might be force_download if you
+            want to update the model.
+
+    Returns:
+        Returns: model_object if include_json is false. (model_object, dict) if include_json is True.
+    """
+    path = Path(path)
+
+    fnames = ["model.pt", "state.pt", "model.json"]
+
+    if all((path / fn).exists() for fn in fnames):
+        fpaths = {fn: path / fn for fn in fnames}
+    else:
+        try:
+            fpaths = {fn: RemoteFile(f"{PRETRAINED_MODELS_BASE_URL}{path}/{fn}", **kwargs).local_path for fn in fnames}
+        except BaseException:
+            raise ValueError(
+                f"No valid model found in {path} or among pre-trained_models at "
+                f"{MATGL_CACHE} or {PRETRAINED_MODELS_BASE_URL}."
+            ) from None
+
+    with open(fpaths["model.json"]) as f:
+        d = json.load(f)
+        modname = d["@module"]
+        classname = d["@class"]
+        model_version = d.get("@model_version", 0)
+        if model_version < MODEL_VERSION:
+            warnings.warn(
+                "Incompatible model version detected! The code will continue to load the model but it is "
+                "recommended that you provide a path to an updated model, increment your @model_version in model.json "
+                "if you are confident that the changes are not problematic, or clear your ~/.matgl cache using "
+                '`python -c "import matgl; matgl.clear_cache()"`',
+                DeprecationWarning,
+                stacklevel=2,
+            )
+        mod = __import__(modname, globals(), locals(), [classname], 0)
+        cls_ = getattr(mod, classname)
+        return cls_.load(path, **kwargs)