+++ matgl/utils/maths.py
@@ -15,22 +15,20 @@ from scipy.optimize import brentq
 from scipy.special import spherical_jn
 from torch import nn
 
-from matgl.config import DataType
-
 CWD = os.path.dirname(os.path.abspath(__file__))
 
 """
 Precomputed Spherical Bessel function roots in a 2D array with dimension [128, 128]. The n-th (0-based index) root of
 order l Spherical Bessel function is the `[l, n]` entry.
 """
-SPHERICAL_BESSEL_ROOTS = np.load(os.path.join(CWD, "sb_roots.npy"))
+SPHERICAL_BESSEL_ROOTS = torch.tensor(np.load(os.path.join(CWD, "sb_roots.npy")))
 
 
 class GaussianExpansion(nn.Module):
     r"""
     Gaussian Radial Expansion.
     The bond distance is expanded to a vector of shape [m],
-    where m is the number of Gaussian basis centers
+    where m is the number of Gaussian basis centers.
     """
 
     def __init__(
@@ -53,16 +51,15 @@ class GaussianExpansion(nn.Module):
                 Width of Gaussian Basis functions
         """
         super().__init__()
-        self.centers = np.linspace(initial, final, num_centers)
-        self.centers = nn.Parameter(torch.tensor(self.centers).float(), requires_grad=False)  # type: ignore
+        self.centers = nn.Parameter(torch.linspace(initial, final, num_centers), requires_grad=False)  # type: ignore
         if width is None:
-            self.width = float(1.0 / np.diff(self.centers).mean())
+            self.width = 1.0 / torch.diff(self.centers).mean()
         else:
             self.width = width
 
     def reset_parameters(self):
         """Reinitialize model parameters."""
-        self.centers = nn.Parameter(self.centers.clone().detach().float(), requires_grad=False)
+        self.centers = nn.Parameter(self.centers, requires_grad=False)
 
     def forward(self, bond_dists):
         """Expand distances.
@@ -88,7 +85,7 @@ def spherical_bessel_roots(max_l: int, max_n: int):
     The calculation is based on the fact that the n-root for l-th
     spherical Bessel function `j_l`, i.e., `z_{j, n}` is in the range
     `[z_{j-1,n}, z_{j-1, n+1}]`. On the other hand we know precisely the
-    roots for j0, i.e., sinc(x)
+    roots for j0, i.e., sinc(x).
 
     Args:
         max_l: max order of spherical bessel function
@@ -111,7 +108,7 @@ def spherical_bessel_roots(max_l: int, max_n: int):
 
 class SphericalBesselFunction:
     """
-    Calculate the spherical Bessel function based on sympy + pytorch implementations
+    Calculate the spherical Bessel function based on sympy + pytorch implementations.
     """
 
     def __init__(self, max_l: int, max_n: int = 5, cutoff: float = 5.0, smooth: bool = False):
@@ -130,8 +127,6 @@ class SphericalBesselFunction:
             self.funcs = self._calculate_smooth_symbolic_funcs()
         else:
             self.funcs = self._calculate_symbolic_funcs()
-
-        self.zeros = torch.tensor(SPHERICAL_BESSEL_ROOTS, dtype=DataType.torch_float)
 
     @lru_cache(maxsize=128)
     def _calculate_symbolic_funcs(self) -> list:
@@ -154,7 +149,7 @@ class SphericalBesselFunction:
     def __call__(self, r):
         """
         Args:
-            r: torch.tensor, distance tensor, 1D
+            r: torch.tensor, distance tensor, 1D.
 
 
         Returns: [n, max_n * max_l] spherical Bessel function results
@@ -169,7 +164,7 @@ class SphericalBesselFunction:
         return torch.t(torch.stack(results))
 
     def _call_sbf(self, r):
-        roots = self.zeros[: self.max_l, : self.max_n]
+        roots = SPHERICAL_BESSEL_ROOTS[: self.max_l, : self.max_n]
 
         results = []
         factor = torch.tensor(sqrt(2.0 / self.cutoff**3))
@@ -186,7 +181,7 @@ class SphericalBesselFunction:
     def rbf_j0(r, cutoff: float = 5.0, max_n: int = 3):
         """
         Spherical Bessel function of order 0, ensuring the function value
-        vanishes at cutoff
+        vanishes at cutoff.
 
         Args:
             r: torch.tensor pytorch tensors
@@ -194,14 +189,14 @@ class SphericalBesselFunction:
             max_n: int max number of basis
         Returns: basis function expansion using first spherical Bessel function
         """
-        n = (torch.arange(1, max_n + 1)).type(dtype=DataType.torch_float)[None, :]
+        n = (torch.arange(1, max_n + 1)).type(dtype=torch.float32)[None, :]
         r = r[:, None]
         return sqrt(2.0 / cutoff) * torch.sin(n * pi / cutoff * r) / r
 
 
 def _y00(theta, phi):
     r"""
-    Spherical Harmonics with `l=m=0`
+    Spherical Harmonics with `l=m=0`.
 
     ..math::
         Y_0^0 = \frac{1}{2} \sqrt{\frac{1}{\pi}}
@@ -222,7 +217,7 @@ def _conjugate(x):
 
 class SphericalHarmonicsFunction:
     """
-    Spherical Harmonics function
+    Spherical Harmonics function.
     """
 
     def __init__(self, max_l: int, use_phi: bool = True):
@@ -230,7 +225,7 @@ class SphericalHarmonicsFunction:
         Args:
             max_l: int, max l (excluding l)
             use_phi: bool, whether to use the polar angle. If not,
-                the function will compute `Y_l^0`
+                the function will compute `Y_l^0`.
         """
         self.max_l = max_l
         self.use_phi = use_phi
@@ -252,17 +247,17 @@ class SphericalHarmonicsFunction:
         """
         Args:
             costheta: Cosine of the azimuthal angle
-            phi: torch.tensor, the polar angle
+            phi: torch.tensor, the polar angle.
 
         Returns: [n, m] spherical harmonic results, where n is the number
             of angles. The column is arranged following
             `[Y_0^0, Y_1^{-1}, Y_1^{0}, Y_1^1, Y_2^{-2}, ...]`
         """
-        costheta = torch.tensor(costheta, dtype=torch.complex64)
-        phi = torch.tensor(phi, dtype=torch.complex64)
-        results = torch.stack([func(costheta, phi) for func in self.funcs], axis=1)
-        results = results.type(dtype=DataType.torch_float)
-        return results
+        # costheta = torch.tensor(costheta, dtype=torch.complex64)
+        # phi = torch.tensor(phi, dtype=torch.complex64)
+        return torch.stack([func(costheta, phi) for func in self.funcs], axis=1)
+        # results = results.type(dtype=DataType.torch_float)
+        # return results
 
 
 def _block_repeat(array, block_size, repeats):
@@ -279,7 +274,7 @@ def _block_repeat(array, block_size, repeats):
 
 def combine_sbf_shf(sbf, shf, max_n: int, max_l: int, use_phi: bool):
     """
-    Combine the spherical Bessel function and the spherical Harmonics function
+    Combine the spherical Bessel function and the spherical Harmonics function.
 
     For the spherical Bessel function, the column is ordered by
         [n=[0, ..., max_n-1], n=[0, ..., max_n-1], ...], max_l blocks,
@@ -327,7 +322,7 @@ def spherical_bessel_smooth(r, cutoff: float = 5.0, max_n: int = 10):
     This is an orthogonal basis with first
     and second derivative at the cutoff
     equals to zero. The function was derived from the order 0 spherical Bessel
-    function, and was expanded by the different zero roots
+    function, and was expanded by the different zero roots.
 
     Ref:
         https://arxiv.org/pdf/1907.02374.pdf
@@ -340,7 +335,7 @@ def spherical_bessel_smooth(r, cutoff: float = 5.0, max_n: int = 10):
     Returns: expanded spherical harmonics with derivatives smooth at boundary
 
     """
-    n = torch.arange(max_n).type(dtype=DataType.torch_float)[None, :]
+    n = torch.arange(max_n).type(dtype=torch.float32)[None, :]
     r = r[:, None]
     fnr = (
         (-1) ** n
@@ -401,7 +396,7 @@ def _get_lambda_func(max_n, cutoff: float = 5.0):
 def get_segment_indices_from_n(ns):
     """
     Get segment indices from number array. For example if
-    ns = [2, 3], then the function will return [0, 0, 1, 1, 1]
+    ns = [2, 3], then the function will return [0, 0, 1, 1, 1].
 
     Args:
         ns: torch.tensor, the number of atoms/bonds array
@@ -411,14 +406,13 @@ def get_segment_indices_from_n(ns):
 
     Returns: segment indices tensor
     """
-    B = ns
-    A = torch.arange(B.size(dim=0))
-    return A.repeat_interleave(B, dim=0)
+    a = torch.arange(ns.size(dim=0))
+    return a.repeat_interleave(ns, dim=0)
 
 
 def get_range_indices_from_n(ns):
     """
-    Give ns = [2, 3], return [0, 1, 0, 1, 2]
+    Give ns = [2, 3], return [0, 1, 0, 1, 2].
 
     Args:
         ns: torch.tensor, the number of atoms/bonds array
@@ -435,25 +429,6 @@ def get_range_indices_from_n(ns):
 
     #    return matrix[mask]
     return torch.masked_select(matrix, mask)
-
-
-# def unsorted_segment_softmax(data, segment_ids, num_segments, weights=None):
-#    """
-#    Unsorted segment softmax with optional weights
-#    Args:
-#        data (tf.Tensor): original data
-#        segment_ids (tf.Tensor): tensor segment ids
-#        num_segments (int): number of segments
-#    Returns: tf.Tensor
-#    """
-#    if weights is None:
-#        weights = torch.ones(1)
-#    segment_max = scatter(data, segment_ids, dim=0, reduce="max")
-#    maxes = torch.gather(segment_max, 0, segment_ids)
-#    data -= maxes
-#    exp = torch.exp(data) * torch.squeeze(weights)
-#    softmax = torch.div(exp, torch.gather(scatter(exp, segment_ids, dim=0, reduce="sum"), 0, segment_ids))
-#    return softmax
 
 
 def repeat_with_n(ns, n):
@@ -473,14 +448,13 @@ def repeat_with_n(ns, n):
 def broadcast_states_to_bonds(g, state_feat):
     """
     Broadcast state attributes of shape [Ns, Nstate] to
-    bond attributes shape [Nb, Nstate]
+    bond attributes shape [Nb, Nstate].
 
     Args:
         g: DGL graph
         state_feat: state_feature
 
     Returns: broadcasted state attributes
-
     """
     return state_feat.repeat((g.num_edges(), 1))
 
@@ -488,7 +462,7 @@ def broadcast_states_to_bonds(g, state_feat):
 def broadcast_states_to_atoms(g, state_feat):
     """
     Broadcast state attributes of shape [Ns, Nstate] to
-    bond attributes shape [Nb, Nstate]
+    bond attributes shape [Nb, Nstate].
 
     Args:
         g: DGL graph
@@ -532,12 +506,11 @@ def unsorted_segment_fraction(data: torch.tensor, segment_ids: torch.tensor, num
         segment_ids (torch.tensor): segment ids
         num_segments (torch.tensor): number of segments
     Returns:
-        data (torch.tensor): data after fraction
+        data (torch.tensor): data after fraction.
     """
     segment_sum = scatter_sum(input_tensor=data, segment_ids=segment_ids, dim=0, num_segments=num_segments)
     sums = torch.gather(segment_sum, 0, segment_ids)
-    data = torch.div(data, sums)
-    return data
+    return torch.div(data, sums)
 
 
 def broadcast(input_tensor: torch.tensor, target_tensor: torch.tensor, dim: int):
@@ -560,5 +533,4 @@ def broadcast(input_tensor: torch.tensor, target_tensor: torch.tensor, dim: int)
         input_tensor = input_tensor.unsqueeze(-1)
     target_shape = list(target_tensor.shape)
     target_shape[dim] = input_tensor.shape[dim]
-    input_tensor = input_tensor.expand(target_shape)
-    return input_tensor
+    return input_tensor.expand(target_shape)