+++ matgl/utils/training.py
@@ -4,6 +4,8 @@ Utils for training MatGL models.
 
 from __future__ import annotations
 
+import math
+
 import dgl
 import numpy as np
 import pytorch_lightning as pl
@@ -44,7 +46,7 @@ class TrainerMixin:
 
     def on_train_epoch_end(self):
         """
-        Step scheduler every epoch
+        Step scheduler every epoch.
         """
         sch = self.lr_schedulers()
         sch.step()
@@ -81,7 +83,7 @@ class TrainerMixin:
 
     def configure_optimizers(self):
         """
-        Configure optimizers
+        Configure optimizers.
         """
         if self.optimizer is None:
             optimizer = torch.optim.Adam(
@@ -109,7 +111,7 @@ class TrainerMixin:
         r"""
         Args:
             *args: Pass-through
-            **kwargs: Pass-through
+            **kwargs: Pass-through.
         """
         super().on_test_model_eval(*args, **kwargs)
         torch.set_grad_enabled(True)
@@ -130,7 +132,7 @@ class TrainerMixin:
 
 class ModelTrainer(TrainerMixin, pl.LightningModule):
     """
-    Trainer for MEGNet and M3GNet models
+    Trainer for MEGNet and M3GNet models.
     """
 
     def __init__(
@@ -156,7 +158,7 @@ class ModelTrainer(TrainerMixin, pl.LightningModule):
             scheduler: scheduler for training
             lr: learning rate for training
             decay_steps: number of steps for decaying learning rate
-            decay_alpha: parameter determines the minimum learning rate
+            decay_alpha: parameter determines the minimum learning rate.
         """
         super().__init__()
 
@@ -186,7 +188,7 @@ class ModelTrainer(TrainerMixin, pl.LightningModule):
         Args:
             g: dgl Graph
             l_g: Line graph
-            state_attr: State attribute
+            state_attr: State attribute.
 
         Returns:
             Model prediction.
@@ -265,7 +267,7 @@ class PotentialTrainer(TrainerMixin, pl.LightningModule):
             scheduler: scheduler for training
             lr: learning rate for training
             decay_steps: number of steps for decaying learning rate
-            decay_alpha: parameter determines the minimum learning rate
+            decay_alpha: parameter determines the minimum learning rate.
         """
         super().__init__()
 
@@ -298,7 +300,7 @@ class PotentialTrainer(TrainerMixin, pl.LightningModule):
         Args:
             g: dgl Graph
             l_g: Line graph
-            state_attr: State attr
+            state_attr: State attr.
 
         Returns:
             energy, force, stress, h
@@ -399,3 +401,21 @@ class PotentialTrainer(TrainerMixin, pl.LightningModule):
             "Force_RMSE": f_rmse,
             "Stress_RMSE": s_rmse,
         }
+
+
+def xavier_init(model: nn.Module) -> None:
+    """Xavier initialization scheme for the model.
+
+    Args:
+        model (nn.Module): The model to be Xavier-initialized.
+    """
+    for name, param in model.named_parameters():
+        if name.endswith(".bias"):
+            param.data.fill_(0)
+        else:
+            if param.dim() < 2:
+                bound = math.sqrt(6) / math.sqrt(param.shape[0] + param.shape[0])
+                param.data.uniform_(-bound, bound)
+            else:
+                bound = math.sqrt(6) / math.sqrt(param.shape[0] + param.shape[1])
+                param.data.uniform_(-bound, bound)